Updated tests to be more general, added Vcsp alive test
