+++ tests/daspeak/conftest.py
@@ -10,22 +10,18 @@ from vericlient.daspeak.exceptions import (
     UnsupportedAudioCodecError,
     UnsupportedSampleRateError,
 )
-from vericlient.environments import Environments, Locations
 from vericlient.exceptions import (
     ServerError,
     UnsupportedMediaTypeError,
 )
 
-from tests.conftest import (
-    eu_production_test_env,
-    eu_sandbox_url,
-    eu_sandobox_test_env,
-    ue_production_url,
-    us_production_test_env,
-    us_production_url,
-    us_sandbox_test_env,
-    us_sandbox_url,
-)
+from tests.conftest import provide_testing_parameters
+
+
+@pytest.fixture(scope="session")
+def service_name():
+    return "daspeak"
+
 
 ####################
 # SERVER RESPONSES #
@@ -259,66 +255,17 @@ def daspeak_server_error_response():
 # PARAMETERS FIXTURES #
 #######################
 
-def _provide_daspeak_parameters(
-        test_environment: str,
-        all_environments: list,
-        mock_option: bool,      # noqa: FBT001
-        endpoint: str,
-        response: dict,
-        status_code: int,
-        exception: Exception,
-    ) -> list:
-    """Provide the parameters necessary for Daspekea testing depending on the test environment.
-
-    Those parameters are:
-    - endpoint: the endpoint to test
-    - response: the response to return
-    - status_code: the status code to return
-    - exception: the exception to raise if any
-
-    The parameters are returned as a list of tuples, each tuple containing the parameters for a specific environment.
-
-    """
-    ue_sandbox = (f"{eu_sandbox_url}/{endpoint}", response, status_code, None, \
-        Environments.SANDBOX.value, Locations.EU.value, exception)
-    ue_production = (f"{ue_production_url}/{endpoint}", response, status_code, None, \
-        Environments.PRODUCTION.value, Locations.EU.value, exception)
-    us_sandbox = (f"{us_sandbox_url}/{endpoint}", response, status_code, None, \
-        Environments.SANDBOX.value, Locations.US.value, exception)
-    us_production = (f"{us_production_url}/{endpoint}", response, status_code, None, \
-        Environments.PRODUCTION.value, Locations.US.value, exception)
-    if mock_option:
-        parameters = [
-            ue_sandbox,
-            ue_production,
-            us_sandbox,
-            us_production,
-            (f"https://custom-daspeak-url.com/{endpoint}", response, status_code, \
-                "https://custom-daspeak-url.com/daspeak/v1", None, None, exception),
-        ]
-    elif test_environment not in all_environments:
-        pytest.fail(f"Invalid environment specified. Use one of {all_environments}")
-    elif test_environment == eu_sandobox_test_env:
-        parameters = [ue_sandbox]
-    elif test_environment == eu_production_test_env:
-        parameters = [ue_production]
-    elif test_environment == us_sandbox_test_env:
-        parameters = [us_sandbox]
-    elif test_environment == us_production_test_env:
-        parameters = [us_production]
-    return parameters
-
-
 @pytest.fixture(scope="session")
 def daspeak_alive_parameters(
         mock_option,
         test_environment,
         all_environments,
         daspeak_alive_response,
+        service_name,
     ) -> list:
-    return _provide_daspeak_parameters(
+    return provide_testing_parameters(
         test_environment, all_environments, mock_option, "daspeak/v1/alive", \
-        daspeak_alive_response, 200, None,
+        daspeak_alive_response, 200, None, service_name,
     )
 
 
@@ -328,10 +275,11 @@ def daspeak_get_models_parameters(
         test_environment,
         all_environments,
         daspeak_get_models_response,
+        service_name,
     )-> list:
-    return _provide_daspeak_parameters(
+    return provide_testing_parameters(
         test_environment, all_environments, mock_option, "daspeak/v1/models", \
-        daspeak_get_models_response, 200, None,
+        daspeak_get_models_response, 200, None, service_name,
     )
 
 
@@ -341,10 +289,11 @@ def daspeak_generate_credential_parameters(
         test_environment,
         all_environments,
         daspeak_generate_credential_response,
+        service_name,
     ) -> list:
-    return _provide_daspeak_parameters(
+    return provide_testing_parameters(
         test_environment, all_environments, mock_option, "daspeak/v1/models/fake-model/credential/wav", \
-        daspeak_generate_credential_response, 200, None,
+        daspeak_generate_credential_response, 200, None, service_name,
     )
 
 
@@ -354,10 +303,11 @@ def daspeak_generate_credential_channels_error_response_parameters(
         test_environment,
         all_environments,
         daspeak_channels_error_response,
+        service_name,
     ):
-    return _provide_daspeak_parameters(
+    return provide_testing_parameters(
         test_environment, all_environments, mock_option, "daspeak/v1/models/fake-model/credential/wav", \
-        daspeak_channels_error_response, 400, TooManyAudioChannelsError,
+        daspeak_channels_error_response, 400, TooManyAudioChannelsError, service_name,
     )
 
 
@@ -367,10 +317,11 @@ def daspeak_generate_credential_sample_rate_error_response_parameters(
         test_environment,
         all_environments,
         daspeak_sample_rate_error_response,
+        service_name,
     ):
-    return _provide_daspeak_parameters(
+    return provide_testing_parameters(
         test_environment, all_environments, mock_option, "daspeak/v1/models/fake-model/credential/wav", \
-        daspeak_sample_rate_error_response, 400, UnsupportedSampleRateError,
+        daspeak_sample_rate_error_response, 400, UnsupportedSampleRateError, service_name,
     )
 
 
@@ -380,10 +331,11 @@ def daspeak_generate_credential_net_speech_duration_error_response_parameters(
         test_environment,
         all_environments,
         daspeak_net_speech_duration_error_response,
+        service_name,
     ):
-    return _provide_daspeak_parameters(
+    return provide_testing_parameters(
         test_environment, all_environments, mock_option, "daspeak/v1/models/fake-model/credential/wav", \
-        daspeak_net_speech_duration_error_response, 400, NetSpeechDurationIsNotEnoughError,
+        daspeak_net_speech_duration_error_response, 400, NetSpeechDurationIsNotEnoughError, service_name,
     )
 
 
@@ -393,10 +345,11 @@ def daspeak_generate_credential_bad_snr_error_response_parameters(
         test_environment,
         all_environments,
         daspeak_snr_error_response,
+        service_name,
     ):
-    return _provide_daspeak_parameters(
+    return provide_testing_parameters(
         test_environment, all_environments, mock_option, "daspeak/v1/models/fake-model/credential/wav", \
-        daspeak_snr_error_response, 400, SignalNoiseRatioError,
+        daspeak_snr_error_response, 400, SignalNoiseRatioError, service_name,
     )
 
 
@@ -406,10 +359,11 @@ def daspeak_generate_credential_audio_too_long_error_response_parameters(
         test_environment,
         all_environments,
         daspeak_long_audio_error_response,
+        service_name,
     ):
-    return _provide_daspeak_parameters(
+    return provide_testing_parameters(
         test_environment, all_environments, mock_option, "daspeak/v1/models/fake-model/credential/wav", \
-        daspeak_long_audio_error_response, 400, AudioDurationTooLongError,
+        daspeak_long_audio_error_response, 400, AudioDurationTooLongError, service_name,
     )
 
 
@@ -419,10 +373,11 @@ def daspeak_generate_credential_codec_error_response_parameters(
         test_environment,
         all_environments,
         daspeak_codec_error_response,
+        service_name,
     ):
-    return _provide_daspeak_parameters(
+    return provide_testing_parameters(
         test_environment, all_environments, mock_option, "daspeak/v1/models/fake-model/credential/wav", \
-        daspeak_codec_error_response, 400, UnsupportedAudioCodecError,
+        daspeak_codec_error_response, 400, UnsupportedAudioCodecError, service_name,
     )
 
 
@@ -432,10 +387,11 @@ def daspeak_generate_credential_insufficient_quality_error_response_parameters(
         test_environment,
         all_environments,
         daspeak_quality_error_response,
+        service_name,
     ):
-    return _provide_daspeak_parameters(
+    return provide_testing_parameters(
         test_environment, all_environments, mock_option, "daspeak/v1/models/fake-model/credential/wav", \
-        daspeak_quality_error_response, 400, InsufficientQualityError,
+        daspeak_quality_error_response, 400, InsufficientQualityError, service_name,
     )
 
 
@@ -445,10 +401,11 @@ def daspeak_generate_credential_invalid_specified_channel_error_response_paramet
         test_environment,
         all_environments,
         daspeak_invalid_specified_channel_error_response,
+        service_name,
     ):
-    return _provide_daspeak_parameters(
+    return provide_testing_parameters(
         test_environment, all_environments, mock_option, "daspeak/v1/models/fake-model/credential/wav", \
-        daspeak_invalid_specified_channel_error_response, 400, InvalidSpecifiedChannelError,
+        daspeak_invalid_specified_channel_error_response, 400, InvalidSpecifiedChannelError, service_name,
     )
 
 
@@ -458,10 +415,11 @@ def daspeak_generate_credential_calibration_not_available_error_response_paramet
         test_environment,
         all_environments,
         daspeak_calibration_not_available_error_response,
+        service_name,
     ):
-    return _provide_daspeak_parameters(
+    return provide_testing_parameters(
         test_environment, all_environments, mock_option, "daspeak/v1/models/fake-model/credential/wav", \
-        daspeak_calibration_not_available_error_response, 400, CalibrationNotAvailableError,
+        daspeak_calibration_not_available_error_response, 400, CalibrationNotAvailableError, service_name,
     )
 
 
@@ -471,10 +429,11 @@ def daspeak_generate_credential_unsupported_media_type_error_response_parameters
         test_environment,
         all_environments,
         daspeak_unsupported_media_type_error_response,
+        service_name,
     ):
-    return _provide_daspeak_parameters(
+    return provide_testing_parameters(
         test_environment, all_environments, mock_option, "daspeak/v1/models/fake-model/credential/wav", \
-        daspeak_unsupported_media_type_error_response, 415, UnsupportedMediaTypeError,
+        daspeak_unsupported_media_type_error_response, 415, UnsupportedMediaTypeError, service_name,
     )
 
 
@@ -484,10 +443,11 @@ def daspeak_server_error_response_parameters(
         test_environment,
         all_environments,
         daspeak_server_error_response,
+        service_name,
     ):
-    return _provide_daspeak_parameters(
+    return provide_testing_parameters(
         test_environment, all_environments, mock_option, "daspeak/v1/models/fake-model/credential/wav", \
-        daspeak_server_error_response, 500, ServerError,
+        daspeak_server_error_response, 500, ServerError, service_name,
     )
 
 
@@ -497,10 +457,11 @@ def daspeak_compare_credential2audio_parameters(
         test_environment,
         all_environments,
         daspeak_compare_credential2audio_response,
+        service_name,
     ) -> list:
-    return _provide_daspeak_parameters(
+    return provide_testing_parameters(
         test_environment, all_environments, mock_option, "daspeak/v1/similarity/credential2wav", \
-        daspeak_compare_credential2audio_response, 200, None,
+        daspeak_compare_credential2audio_response, 200, None, service_name,
     )
 
 
@@ -510,10 +471,11 @@ def daspeak_compare_audio2audio_parameters(
         test_environment,
         all_environments,
         daspeak_compare_audio2audio_response,
+        service_name,
     ) -> list:
-    return _provide_daspeak_parameters(
+    return provide_testing_parameters(
         test_environment, all_environments, mock_option, "daspeak/v1/similarity/wav2wav", \
-        daspeak_compare_audio2audio_response, 200, None,
+        daspeak_compare_audio2audio_response, 200, None, service_name,
     )
 
 
@@ -523,10 +485,11 @@ def daspeak_compare_credential2credential_parameters(
         test_environment,
         all_environments,
         daspeak_compare_credential2credential_response,
+        service_name,
     ) -> list:
-    return _provide_daspeak_parameters(
+    return provide_testing_parameters(
         test_environment, all_environments, mock_option, "daspeak/v1/similarity/credential2credential", \
-        daspeak_compare_credential2credential_response, 200, None,
+        daspeak_compare_credential2credential_response, 200, None, service_name,
     )
 
 
@@ -536,10 +499,11 @@ def daspeak_compare_audio2credentials_parameters(
         test_environment,
         all_environments,
         daspeak_compare_audio2credentials_response,
+        service_name,
     ) -> list:
-    return _provide_daspeak_parameters(
+    return provide_testing_parameters(
         test_environment, all_environments, mock_option, "daspeak/v1/identification/wav2credentials", \
-        daspeak_compare_audio2credentials_response, 200, None,
+        daspeak_compare_audio2credentials_response, 200, None, service_name,
     )
 
 
@@ -549,77 +513,9 @@ def daspeak_compare_credential2credentials_parameters(
         test_environment,
         all_environments,
         daspeak_compare_credential2credentials_response,
+        service_name,
     ) -> list:
-    return _provide_daspeak_parameters(
+    return provide_testing_parameters(
         test_environment, all_environments, mock_option, "daspeak/v1/identification/credential2credentials", \
-        daspeak_compare_credential2credentials_response, 200, None,
+        daspeak_compare_credential2credentials_response, 200, None, service_name,
     )
-
-
-######################
-# RESOURCES FIXTURES #
-######################
-
-@pytest.fixture(scope="session")
-def audio_file_path() -> str:
-    return "tests/daspeak/resources/audio.wav"
-
-
-@pytest.fixture(scope="session")
-def audio_file() -> bytes:
-    with open("tests/daspeak/resources/audio.wav", "rb") as f:
-        return f.read()
-
-
-@pytest.fixture(scope="session")
-def audio_spoof_file() -> bytes:
-    with open("tests/daspeak/resources/audio_spoof.wav", "rb") as f:
-        return f.read()
-
-
-@pytest.fixture(scope="session")
-def audio_not_enough_speech_file() -> bytes:
-    with open("tests/daspeak/resources/audio_not_enough_speech.wav", "rb") as f:
-        return f.read()
-
-
-@pytest.fixture(scope="session")
-def audio_bad_snr_file() -> bytes:
-    with open("tests/daspeak/resources/audio_bad_snr.wav", "rb") as f:
-        return f.read()
-
-
-@pytest.fixture(scope="session")
-def audio_insufficient_quality_file() -> bytes:
-    with open("tests/daspeak/resources/audio_insufficient_quality.wav", "rb") as f:
-        return f.read()
-
-
-@pytest.fixture(scope="session")
-def audio_too_many_channels_file() -> bytes:
-    with open("tests/daspeak/resources/audio_too_many_channels.wav", "rb") as f:
-        return f.read()
-
-
-@pytest.fixture(scope="session")
-def audio_invalid_sample_rate_file() -> bytes:
-    with open("tests/daspeak/resources/audio_invalid_sample_rate.wav", "rb") as f:
-        return f.read()
-
-
-@pytest.fixture(scope="session")
-def audio_too_long_file() -> bytes:
-    with open("tests/daspeak/resources/audio_too_long.wav", "rb") as f:
-        return f.read()
-
-
-@pytest.fixture(scope="session")
-def audio_codec_error_file() -> bytes:
-    with open("tests/daspeak/resources/audio_codec_error.wav", "rb") as f:
-        return f.read()
-
-
-@pytest.fixture(scope="session")
-def audio_unsupported_file() -> bytes:
-    with open("tests/__init__.py", "rb") as f:
-        return f.read()
